fixed bbgun.js firing dom events, bbgun initialization in bb.js
When a BBGun is created,
it asks bb.js to create a new dom element.
If the new element had an event,
then it needed to be added to the BBGun,
but was not being done. This is now fixed.

The dom is now attached to BBGun, by bb.js,
at the earliest possible time, right after it has been
created.

The other bug fixed is that when dom events were
fired, they did not fire the events inside BBGun.
This is because event name was accessed using
the wrong property.